Added DynamicOctree and IShape
IShape is an interface for shape-like classes whose volume can be enclosed by an axis-aligned bounding box.

The DynamicOctree models a dynamic octree which automatically branches and merges its nodes when IShapes are inserted or removed from the tree. A DynamicOctree has three configurable elements: the looseness of the tree, minimum width of a node, and the maximum number of children in a node.